slow down (50%) the rate at which Z axis descends during final stage of depth probe
in an effort to make more repeatable cuts after changing bits.
I've noticed that successive cuts on the same piece with different tools are at slightly different depths. I run the depth probe sequence after each tool change, of course. Perhaps slowing down the final latch phase could improve repeatability.
The other possibility is that the surface of the depth probe switch is not parallel to the U rail it's embedded in. This would cause different depth readings depending on where the cutter head contacts the switch surface.